\encoding{latin1}
\name{center}
\alias{center}
\title{Centering of matrices}
\description{
  Centers a matrix by subtracting its coloumn means.
}
\usage{
center(Y)
}
\arguments{
  \item{Y}{a matrix.}
}
%\details{
%}
\value{
  A matrix.
}
\author{Øyvind Langsrud}
\seealso{ ~~objects to See Also as \code{\link{help}}, ~~~ }
\examples{
}
\keyword{ ~kwd1 }% at least one, from doc/KEYWORDS
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
