% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{adjust}
\alias{adjust}
\title{Adjust a predictor matrix for the presence of another matrix}
\usage{
adjust(X, Y)
}
\arguments{
\item{X}{matrix.  The matrix to be adjusted.}

\item{Y}{matrix.  The matrix to be adjusted for.}
}
\value{
A matrix with an orthogonal basis for the adjusted predictor matrix.
}
\description{
\code{adjust} adjusts a predictor matrix \eqn{X} for the presence of another
predictor matrix \eqn{Y}, by orthogonalizing \eqn{X} against \eqn{Y}.
}
\details{
The function can handle rank deficient matrices.
}
\author{
Øyvind Langsrud
}
\keyword{internal}
\keyword{models}
