\encoding{latin1}
\name{ffmanova}
\alias{ffmanova}
\title{Fifty-fifty MANOVA}
\description{
  Calculate a 50-50 Manova.  FIXME.
}
\usage{
ffmanova(formel, Y, stand = 1, nSim = 0)
}
\arguments{
  \item{formel}{formula.}
  \item{Y}{matrix.  Response data(?)}
  \item{stand}{numeric.  Whether \code{Y} should be standardized. FIXME:
  should maybe be a logical}
  \item{nSim}{nonnegative integer.  The number of simulations to use in
    the rotation tests(?)}
}
\details{
  FIXME: Fyll ut.
}
\value{
  ~Describe the value returned
  If it is a LIST, use
  \item{comp1 }{Description of 'comp1'}
  \item{comp2 }{Description of 'comp2'}
  ...
}
\references{ ~put references to the literature/web site here ~ }
\author{Øyvind Langsrud}
\note{ ~~further notes~~ 

 ~Make other sections like Warning with \section{Warning }{....} ~
}
\seealso{ ~~objects to See Also as \code{\link{help}}, ~~~ }
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.
}
\keyword{ ~kwd1 }% at least one, from doc/KEYWORDS
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
