% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffmanovatest.R
\name{my_pValueF}
\alias{my_pValueF}
\title{F-test p-values}
\usage{
my_pValueF(f, ny1, ny2)
}
\arguments{
\item{f}{The \eqn{F} value}

\item{ny1}{The numerator df's}

\item{ny2}{The denominator df's}
}
\value{
A \eqn{p}-value.
}
\description{
This is simply a wrapper around \code{\link{pf}}: \code{my_pValueF(f, ny1,
ny2)} is equivalent to \code{pf(f, ny1, ny2, lower.tail = FALSE)}.
}
\seealso{
\code{\link{pf}}
}
\author{
Øyvind Langsrud and Bjørn-Helge Mevik
}
\keyword{htest}
\keyword{internal}
