% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffmanova_div.R
\name{norm}
\alias{norm}
\title{Matrix norm.}
\usage{
norm(X)
}
\arguments{
\item{X}{a numeric matrix.}
}
\value{
The largest singular value of \code{X}.
}
\description{
\code{norm(X)} returns the largest singular value of \code{X}; it is
equivalent to \code{svd(X, nu = 0, nv = 0)$d[1]}.
}
\seealso{
\code{\link{svd}}
}
\author{
Øyvind Langsrud and Bjørn-Helge Mevik
}
\keyword{algebra}
\keyword{array}
\keyword{internal}
