% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_Obj.R
\name{orth_D}
\alias{orth_D}
\title{Making adjusted design matrix data}
\usage{
orth_D(D, model, method)
}
\arguments{
\item{D}{A list containing a regressor matrix for each model term}

\item{model}{The model coded as a matrix}

\item{method}{Either \code{"test"} or \code{"om"}}
}
\value{
An adjusted version of \code{D} is returned.
}
\description{
The function takes the output from \code{modelData} as input and calculates
adjusted data
}
\details{
The \code{"test"} method adjusts data according to Type II* sums of squares.
This is an extension of the traditional Type II method. The \code{"om"}
method orthogonalises terms according to the model hierarchy. The result is
a non-overparameterised representation of the model.
}
\author{
Øyvind Langsrud and Bjørn-Helge Mevik
}
\keyword{design}
\keyword{internal}
\keyword{models}
