% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{print.ffmanova}
\alias{print.ffmanova}
\title{Print method for ffmanova}
\usage{
\method{print}{ffmanova}(x, digits = max(getOption("digits") - 3, 3), ...)
}
\arguments{
\item{x}{\code{"ffmanova"} object.  Typically created by
\code{\link{ffmanova}}.}

\item{digits}{positive integer.  Minimum number of significant digits to be
used for printing most numbers.}

\item{\dots}{further arguments sent to the underlying
\code{\link{printCoefmat}}.}
}
\value{
Invisibly returns the original object.
}
\description{
Print method for objects of class \code{"ffmanova"}.  It prints an ANOVA
table.
}
\details{
The function constructs an anova table, and prints it using
\code{\link{printCoefmat}} with tailored arguments.
}
\seealso{
\code{\link{ffmanova}}, \code{\link{printCoefmat}}
}
\author{
Bjørn-Helge Mevik
}
\keyword{internal}
\keyword{print}
